requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
colorama==0.4.4
Flask==2.1.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.20.0
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.0
